--- app/src/main/res/layout/fragment_photo.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/black"
-    tools:context=".cameraUtils.PhotoFragment">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <ImageButton
-            android:id="@+id/flash_button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="10sp"
-            android:background="?attr/selectableItemBackground"
-            android:backgroundTint="@color/transparent"
-            android:contentDescription="@string/flash_toggle"
-            android:elevation="2dp"
-            android:padding="15sp"
-            android:src="@drawable/flash_off_black_24dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-
-        <androidx.camera.view.PreviewView
-            android:id="@+id/photo_view_finder"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintDimensionRatio="9:16"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent">
-
-        </androidx.camera.view.PreviewView>
-
-        <ImageButton
-            android:id="@+id/shutter_button"
-            android:layout_width="64dp"
-            android:layout_height="64dp"
-            android:layout_marginBottom="10sp"
-            android:background="?attr/selectableItemBackground"
-            android:backgroundTint="@color/primary_500"
-            android:contentDescription="@string/start_capture"
-            android:elevation="2dp"
-            android:src="@drawable/capture_white_48dp"
-            app:layout_constraintBottom_toBottomOf="@+id/photo_view_finder"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</FrameLayout>
