--- app/src/main/res/layout/fragment_video.xml
@@ -1,106 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/black"
-    tools:context=".cameraUtils.VideoFragment">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <ImageButton
-            android:id="@+id/flash_button_v"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="10sp"
-            android:background="?attr/selectableItemBackground"
-            android:backgroundTint="@color/transparent"
-            android:contentDescription="@string/flash_toggle"
-            android:elevation="2dp"
-            android:padding="15sp"
-            android:src="@drawable/flash_off_black_24dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="@+id/video_view_finder" />
-
-        <TextView
-            android:id="@+id/videoTimer"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="10sp"
-            android:padding="15sp"
-            android:text="@string/video_timer"
-            android:textColor="@color/white"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="@+id/video_view_finder" />
-
-        <Button
-            android:id="@+id/quality_selector"
-            style="@style/Widget.Material3.Button.TextButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="10sp"
-            android:layout_marginEnd="5sp"
-            android:background="?attr/selectableItemBackground"
-            android:backgroundTint="@color/transparent"
-            android:contentDescription="@string/video_quality"
-            android:includeFontPadding="false"
-            android:minWidth="0dp"
-            android:minHeight="0dp"
-            android:padding="15sp"
-            android:text="@string/q720p"
-            android:textColor="@color/white"
-            app:layout_constraintEnd_toStartOf="@+id/fps_selector"
-            app:layout_constraintTop_toTopOf="@+id/video_view_finder" />
-
-        <Button
-            android:id="@+id/fps_selector"
-            style="@style/Widget.Material3.Button.TextButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="10sp"
-            android:layout_marginEnd="5sp"
-            android:background="?attr/selectableItemBackground"
-            android:backgroundTint="@color/transparent"
-            android:contentDescription="@string/video_fps"
-            android:elevation="2dp"
-            android:minWidth="0dp"
-            android:minHeight="0dp"
-            android:padding="15sp"
-            android:text="@string/fps30"
-            android:textColor="@color/white"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="@+id/video_view_finder" />
-
-        <ImageButton
-            android:id="@+id/shutter_button_v"
-            android:layout_width="64dp"
-            android:layout_height="64dp"
-            android:layout_marginBottom="10sp"
-            android:background="?attr/selectableItemBackground"
-            android:backgroundTint="@color/primary_500"
-            android:contentDescription="@string/start_capture"
-            android:elevation="2dp"
-            android:src="@drawable/record_start_black_24dp"
-            app:layout_constraintBottom_toBottomOf="@+id/video_view_finder"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent" />
-
-        <androidx.camera.view.PreviewView
-            android:id="@+id/video_view_finder"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            android:layout_marginTop="5sp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintDimensionRatio="9:16"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent">
-
-        </androidx.camera.view.PreviewView>
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</FrameLayout>
