+++ app/src/main/res/values/strings.xml
@@ -10,14 +10,11 @@
     <string name="video">Video</string>
     <string name="q1080p">1080p</string>
     <string name="q720p">720p</string>
-    <string name="fps30">30 FPS</string>
-    <string name="fps60">60 FPS</string>
+    <string name="q480p">480p</string>
     <string name="video_quality">Video Quality</string>
-    <string name="video_fps">Video FPS</string>
-    <string name="take_photo">Take Photo</string>
+    <string name="switch_camera">Switch Camera</string>
     <string name="flash_toggle">Flash</string>
     <string name="start_capture">Start Capture</string>
-    <string name="stop_capture">Stop Capture</string>
     <string name="video_error">Internal Error, could not save video</string>
     <string name="video_timer">00:00</string>
     <string name="permission_denied">Permission request denied</string>
@@ -86,8 +83,5 @@
     <string name="root_check_title">Root Check</string>
     <string name="root_check_subtitle">It is not recommended to use this app on a rooted phone. The app will now close.</string>
     <!-- TODO: Remove or change this placeholder text -->
-    <string name="image_capture_fragment">Image Capture</string>
-    <string name="video_capture_fragment">Video Capture</string>
     <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
